backendV2/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
httpx>=0.25.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0